requirements.txt
customtkinter>=5.2.0
mysql-connector-python>=8.3.0
bcrypt>=4.1.0
Pillow>=10.0.0
tkcalendar>=1.6.1
openpyxl>=3.1.0
reportlab>=4.0.0
reportlab>=4.0.0
